requirements.txt
numpy 
rasterio
pandas
matplotlib
scikit-learn
torch
torchvision
mlflow
pydotenv
scipy
ujson
tqdm
python-dotenv